refactor : 유닛테스트 실패 오류 수정
package name을 변경하여 유닛테스트 실패 오류를 수정함
