+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ mybatis.type-aliases-package=flab.resellPlatform.domain
 logging.level.flab.resellPlatform.userEntity.repository=trace
 
 # MySQL8
-spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:${MYSQL_PORT:3306}/resell_platform_v2
+spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:${MYSQL_PORT:3306}/${DB_SCHEMA:v_1}
 spring.datasource.username=root
 spring.datasource.password=root
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
